11.10.2024

+++ data/src/main/java/ru/madbrains/data/local/entity/AddressDoorEntity.kt
@@ -25,7 +25,7 @@ data class AddressDoorEntity(
     val icon: String,
 
     @ColumnInfo(name = "domophoneId")
-    val domophoneId: Int,
+    val domophoneId: Long,
 
     @ColumnInfo(name = "doorId")
     val doorId: Int?,
