+++ domain/src/main/java/ru/madbrains/domain/model/request/OpenDoorRequest.kt
@@ -3,6 +3,6 @@ package ru.madbrains.domain.model.request
 import com.squareup.moshi.Json
 
 data class OpenDoorRequest(
-    @Json(name = "domophoneId") val domophoneId: Int,
+    @Json(name = "domophoneId") val domophoneId: Long,
     @Json(name = "doorId") val doorId: Int?
 )
